pyproject.toml
[tool.poetry]
name = "chat"
version = "0.1.0"
description = ""
authors = ["SarlosCainz <m-takuj@sarlos.jp>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.95.0"
uvicorn = "^0.21.1"
mangum = "^0.17.0"
fastapi-utils = "^0.2.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
pydantic = {extras = ["email"], version = "^1.10.6"}
requests = "^2.28.2"
python-multipart = "^0.0.6"


[tool.poetry.group.dev.dependencies]
boto3 = "^1.26.94"
aws-cdk-lib = "^2.69.0"
pytest = "^7.2.2"
constructs = "^10.1.281"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

